# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "@openapi:info",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/equipment": {
      "get": {
        "description": "List all equipment in reef-pi.",
        "tags": [
          "Equipment"
        ],
        "summary": "List all equipment.",
        "operationId": "equipmentList",
        "responses": {
          "200": {
            "description": "equipment",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/equipment"
              }
            }
          }
        }
      },
      "put": {
        "description": "Create a new equipment.",
        "tags": [
          "Equipment"
        ],
        "summary": "Create an equipment.",
        "operationId": "equipmentCreate",
        "parameters": [
          {
            "description": "The equipment to create",
            "name": "equipment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/equipment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "Update an existing equipment.",
        "tags": [
          "Equipment"
        ],
        "summary": "Update an equipment.",
        "operationId": "equipmentUpdate",
        "parameters": [
          {
            "description": "The Id of the equipment to update",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The equipment to update",
            "name": "equipment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/equipment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/equipment/{id}": {
      "get": {
        "description": "Get an existing equipment by id.",
        "tags": [
          "Equipment"
        ],
        "summary": "Get an equipment by id.",
        "operationId": "equipmentGet",
        "parameters": [
          {
            "description": "The Id of the equipment",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/equipment"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "Delete an existing equipment.",
        "tags": [
          "Equipment"
        ],
        "summary": "Delete an equipment.",
        "operationId": "equipmentDelete",
        "parameters": [
          {
            "description": "The Id of the equipment to delete",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/equipment/{id}/control": {
      "post": {
        "description": "Control an equipment.",
        "tags": [
          "Equipment"
        ],
        "summary": "Control an equipment.",
        "operationId": "equipmentControl",
        "parameters": [
          {
            "description": "The Id of the equipment to control",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The action to take",
            "name": "action",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/equipmentAction"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/inlets": {
      "get": {
        "description": "List all inlets in reef-pi.",
        "tags": [
          "Inlet"
        ],
        "summary": "List all Inlets.",
        "operationId": "inletList",
        "responses": {
          "200": {
            "description": "inlet",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/inlet"
              }
            }
          }
        }
      },
      "put": {
        "description": "Create a new Inlet.",
        "tags": [
          "Inlet"
        ],
        "summary": "Create an Inlet.",
        "operationId": "inletCreate",
        "parameters": [
          {
            "description": "The inlet to create",
            "name": "inlet",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/inlet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/inlets/{id}": {
      "get": {
        "description": "Get an existing Inlet.",
        "tags": [
          "Inlet"
        ],
        "summary": "Get an Inlet by id.",
        "operationId": "inletGet",
        "parameters": [
          {
            "description": "The Id of the inlet",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/inlet"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "Update an existing Inlet.",
        "tags": [
          "Inlet"
        ],
        "summary": "Update an Inlet.",
        "operationId": "inletUpdate",
        "parameters": [
          {
            "description": "The Id of the inlet to update",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The inlet to update",
            "name": "inlet",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/inlet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "Delete an existing Inlet.",
        "tags": [
          "Inlet"
        ],
        "summary": "Delete an Inlet.",
        "operationId": "inletDelete",
        "parameters": [
          {
            "description": "The Id of the inlet to delete",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/inlets/{id}/read": {
      "post": {
        "description": "Read an Inlet.",
        "tags": [
          "Inlet"
        ],
        "summary": "Read an Inlet.",
        "operationId": "inletRead",
        "parameters": [
          {
            "description": "The Id of the inlet to read",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "equipment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "on": {
          "type": "boolean",
          "x-go-name": "On"
        },
        "outlet": {
          "type": "string",
          "x-go-name": "Outlet"
        }
      },
      "x-go-name": "Equipment",
      "x-go-package": "github.com/reef-pi/reef-pi/controller/modules/equipment"
    },
    "equipmentAction": {
      "type": "object",
      "properties": {
        "on": {
          "type": "boolean",
          "x-go-name": "On"
        }
      },
      "x-go-name": "EquipmentAction",
      "x-go-package": "github.com/reef-pi/reef-pi/controller/modules/equipment"
    },
...
    "inlet": {
      "type": "object",
      "properties": {
        "driver": {
          "type": "string",
          "x-go-name": "Driver"
        },
        "equipment": {
          "type": "string",
          "x-go-name": "Equipment"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "pin": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Pin"
        },
        "reverse": {
          "type": "boolean",
          "x-go-name": "Reverse"
        }
      },
      "x-go-name": "Inlet",
      "x-go-package": "github.com/reef-pi/reef-pi/controller/connectors"
    },
...
  }
}